refact: 컴포넌트 사용을 통한 코드 간소화
