--- gible/src/components/signup/Title.js
@@ -1,8 +1,0 @@
-import React from "react";
-import styled from "styled-components";
-
-const Title = () => {
-  return <div></div>;
-};
-
-export default Title;
