+++ gible/src/pages/signup/SignUp01.js
@@ -1,38 +1,14 @@
 import React from "react";
 import styled from "styled-components";
-import { ReactComponent as chevron } from "@/assets/chevron-right.svg";
+import SignUpTitle from "@/components/signup/SignUpTitle";
+import SignUpStepBox from "@/components/signup/SignUpStepBox";
 
 const SignUp01 = () => {
   return (
     <Box>
       <Center>
-        <TitleBox>
-          <Title>회원가입</Title>
-          <SubTitle>
-            <span style={{ color: "red" }}>한 장의 기부</span>가 누군가에겐
-            <span style={{ color: "red" }}> 큰 힘</span>이 됩니다.
-          </SubTitle>
-          <Description>
-            기블 홈페이지 회원이 되신다면 더욱 편리하게 서비스 이용이
-            가능합니다.
-          </Description>
-        </TitleBox>
-        <StepBox>
-          <StepTextBox style={{ color: "red" }}>
-            <StepText>STEP 01</StepText>
-            <StepDescription>약관동의</StepDescription>
-          </StepTextBox>
-          <Chevron />
-          <StepTextBox>
-            <StepText>STEP 02</StepText>
-            <StepDescription>기본정보입력</StepDescription>
-          </StepTextBox>
-          <Chevron />
-          <StepTextBox>
-            <StepText>STEP 03</StepText>
-            <StepDescription>가입완료</StepDescription>
-          </StepTextBox>
-        </StepBox>
+        <SignUpTitle />
+        <SignUpStepBox number={1} />
         <CheckBoxBox>
           <CheckBoxTextBox>
             <CheckBox type="checkbox" />
@@ -98,82 +74,6 @@ const Box = styled.div`
 `;
 
 const Center = styled.div``;
-
-const TitleBox = styled.div`
-  margin-bottom: 50px;
-`;
-
-const Title = styled.div`
-  font-size: 26px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: var(--24, 34px); /* 130.769% */
-  letter-spacing: -0.65px;
-  margin-bottom: 50px;
-`;
-
-const SubTitle = styled.div`
-  font-size: 40px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: var(--36, 48px); /* 120% */
-  letter-spacing: -1px;
-  margin-bottom: 8px;
-`;
-
-const Description = styled.div`
-  color: #767676;
-  font-size: 18px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: var(--18, 26px); /* 144.444% */
-  letter-spacing: -0.45px;
-`;
-
-const StepBox = styled.div`
-  color: #767676;
-  border-radius: 6px;
-  border: 1px solid #dbdbdb;
-  background: #fef3f4;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const StepTextBox = styled.div`
-  display: flex;
-  width: 388px;
-  height: 120px;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 5px;
-  flex-shrink: 0;
-`;
-
-const StepText = styled.div`
-  text-align: center;
-  font-size: 24px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: var(--24, 34px); /* 141.667% */
-  letter-spacing: -0.6px;
-`;
-
-const StepDescription = styled.div`
-  text-align: center;
-  font-size: 18px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: var(--18, 26px); /* 144.444% */
-  letter-spacing: -0.45px;
-`;
-
-const Chevron = styled(chevron)`
-  width: var(--26, 36px);
-  height: var(--26, 36px);
-  flex-shrink: 0;
-`;
 
 const CheckBoxBox = styled.div`
   padding: 20px 0px;
