+++ gible/src/pages/signup/SignUp03.js
@@ -1,31 +1,16 @@
 import React from "react";
 import styled from "styled-components";
-import { ReactComponent as chevron } from "@/assets/chevron-right.svg";
+import SignUpTitle from "@/components/signup/SignUpTitle";
+import SignUpStepBox from "@/components/signup/SignUpStepBox";
+
 import { ReactComponent as check } from "@/assets/circle-check-solid.svg";
 
 const SignUp03 = () => {
   return (
     <Box>
       <Center>
-        <TitleBox>
-          <Title>회원가입</Title>
-        </TitleBox>
-        <StepBox>
-          <StepTextBox>
-            <StepText>STEP 01</StepText>
-            <StepDescription>약관동의</StepDescription>
-          </StepTextBox>
-          <Chevron />
-          <StepTextBox>
-            <StepText>STEP 02</StepText>
-            <StepDescription>기본정보입력</StepDescription>
-          </StepTextBox>
-          <Chevron />
-          <StepTextBox style={{ color: "red" }}>
-            <StepText>STEP 03</StepText>
-            <StepDescription>가입완료</StepDescription>
-          </StepTextBox>
-        </StepBox>
+        <SignUpTitle />
+        <SignUpStepBox number={3} />
       </Center>
       <Check />
       <WelcomeTitle>
@@ -56,65 +41,6 @@ const Box = styled.div`
 `;
 
 const Center = styled.div``;
-
-const TitleBox = styled.div`
-  margin-bottom: 50px;
-`;
-
-const Title = styled.div`
-  font-size: 26px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: var(--24, 34px); /* 130.769% */
-  letter-spacing: -0.65px;
-  margin-bottom: 50px;
-`;
-
-const StepBox = styled.div`
-  color: #767676;
-  border-radius: 6px;
-  border: 1px solid #dbdbdb;
-  background: #fef3f4;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 58px;
-`;
-
-const StepTextBox = styled.div`
-  display: flex;
-  width: 388px;
-  height: 120px;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 5px;
-  flex-shrink: 0;
-`;
-
-const StepText = styled.div`
-  text-align: center;
-  font-size: 24px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: var(--24, 34px); /* 141.667% */
-  letter-spacing: -0.6px;
-`;
-
-const StepDescription = styled.div`
-  text-align: center;
-  font-size: 18px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: var(--18, 26px); /* 144.444% */
-  letter-spacing: -0.45px;
-`;
-
-const Chevron = styled(chevron)`
-  width: var(--26, 36px);
-  height: var(--26, 36px);
-  flex-shrink: 0;
-`;
 
 const Check = styled(check)`
   width: 56px;
